backend/dcsBackend/sessionManagementApi/requirements.txt
Flask==2.0.2
Flask-Cors==3.10.9
Flask-JWT-Extended==4.3.1
Flask-MySQLdb==1.0.0
python-dotenv==0.19.2
requests==2.26.0
